rename and split dts (#14)
* rename and split dts

The convention for sharing common features for different platforms is to
use .dtsi for describing commonalities, then .dts files for concrete
platforms.  Split the .dts file into the .dtsi, but move the cpus,
fcram, and bootargs into the machine specific .dts file.

This should enable a simple nintendo3ds_ktr.dts file to be added.

To build:
$ ARCH=arm CROSS_COMPILE=arm-linux-gnueabi- make \
  arch/arm/boot/dts/nintendo3ds_ctr.dtb

To observe what the combined DT looks like:
$ ./scripts/dtc/dtc -O dts arch/arm/boot/dts/nintendo3ds_ctr.dtb

Fixes #12

* add additional level for compatible
* add ktr dts for now

Still depends on #1 and #2 be resolved in the bootloader before we can
use this.

* fixup compatible strings
* Add codenames to dts files

Added codenames from:
https://nindb.net/3ds/hardware.html

* add todo
* add 2ds xl to ktr dts

Signed-off-by: Nick Desaulniers <nick.desaulniers@gmail.com>